# Add MTE-1549 [v119] Optimize XCUITest wait methods