--- a/L10nSnapshotTests/L10nBaseSnapshotTests.swift
+++ b/L10nSnapshotTests/L10nBaseSnapshotTests.swift
@@ -47,7 +47,7 @@ class L10nBaseSnapshotTests: XCTestCase {
     }
 
     // is up to 25x more performant than the above waitForExistence method
-    func mozWaitForElementToExist(element: XCUIElement, timeout: TimeInterval? = TIMEOUT) {
+    func mozWaitForElementToExist(_ element: XCUIElement, timeout: TimeInterval? = TIMEOUT) {
         let startTime = Date()
 
         while !element.exists {
@@ -76,7 +76,7 @@ class L10nBaseSnapshotTests: XCTestCase {
         waitFor(element, with: "exists != true", timeout: timeoutValue, file: file, line: line)
     }
     // is up to 25x more performant than the above waitForNoExistence method
-    func mozWaitForElementToNotExist(element: XCUIElement, timeout: TimeInterval? = TIMEOUT) {
+    func mozWaitForElementToNotExist(_ element: XCUIElement, timeout: TimeInterval? = TIMEOUT) {
         let startTime = Date()
 
         while element.exists {