--- a/L10nSnapshotTests/L10nSuite1SnapshotTests.swift
+++ b/L10nSnapshotTests/L10nSuite1SnapshotTests.swift
@@ -27,33 +27,33 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
     }
 
     func testIntro() {
-        waitForExistence(app.scrollViews.staticTexts["\(rootA11yId)TitleLabel"], timeout: 15)
-        waitForExistence(app.scrollViews.staticTexts["\(rootA11yId)DescriptionLabel"], timeout: 15)
+        mozWaitForElementToExist(app.scrollViews.staticTexts["\(rootA11yId)TitleLabel"], timeout: 15)
+        mozWaitForElementToExist(app.scrollViews.staticTexts["\(rootA11yId)DescriptionLabel"], timeout: 15)
         snapshot("Onboarding-1")
 
         // Swipe to the second screen
         app.buttons["\(rootA11yId)SecondaryButton"].tap()
         currentScreen += 1
-        waitForExistence(app.scrollViews.staticTexts["\(rootA11yId)TitleLabel"], timeout: 15)
-        waitForExistence(app.scrollViews.staticTexts["\(rootA11yId)DescriptionLabel"], timeout: 15)
-        waitForExistence(app.buttons["\(rootA11yId)PrimaryButton"])
-        waitForExistence(app.buttons["\(rootA11yId)SecondaryButton"])
+        mozWaitForElementToExist(app.scrollViews.staticTexts["\(rootA11yId)TitleLabel"], timeout: 15)
+        mozWaitForElementToExist(app.scrollViews.staticTexts["\(rootA11yId)DescriptionLabel"], timeout: 15)
+        mozWaitForElementToExist(app.buttons["\(rootA11yId)PrimaryButton"])
+        mozWaitForElementToExist(app.buttons["\(rootA11yId)SecondaryButton"])
         snapshot("Onboarding-2")
 
         // Swipe to the third screen
         app.buttons["\(rootA11yId)SecondaryButton"].tap()
         currentScreen += 1
-        waitForExistence(app.scrollViews.staticTexts["\(rootA11yId)TitleLabel"], timeout: 15)
-        waitForExistence(app.scrollViews.staticTexts["\(rootA11yId)DescriptionLabel"], timeout: 15)
-        waitForExistence(app.buttons["\(rootA11yId)PrimaryButton"])
-        waitForExistence(app.buttons["\(rootA11yId)SecondaryButton"])
+        mozWaitForElementToExist(app.scrollViews.staticTexts["\(rootA11yId)TitleLabel"], timeout: 15)
+        mozWaitForElementToExist(app.scrollViews.staticTexts["\(rootA11yId)DescriptionLabel"], timeout: 15)
+        mozWaitForElementToExist(app.buttons["\(rootA11yId)PrimaryButton"])
+        mozWaitForElementToExist(app.buttons["\(rootA11yId)SecondaryButton"])
         snapshot("Onboarding-3")
 
         // Swipe to the Homescreen
         app.buttons["\(rootA11yId)SecondaryButton"].tap()
         currentScreen += 1
-        waitForExistence(app.textFields["url"])
-        waitForExistence(app.webViews["contentView"])
+        mozWaitForElementToExist(app.textFields["url"])
+        mozWaitForElementToExist(app.webViews["contentView"])
         snapshot("Homescreen-first-visit")
     }
 
@@ -69,8 +69,8 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 //
 //        // Link
 //        navigator.openURL("http://wikipedia.org")
-//        waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
-//        waitForExistence(app.webViews.element(boundBy: 0).links.element(boundBy: 0), timeout: 5)
+//        mozWaitForElementToNotExist(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+//        mozWaitForElementToExist(app.webViews.element(boundBy: 0).links.element(boundBy: 0), timeout: 5)
 //        navigator.goto(WebLinkContextMenu)
 //        drag()
 //        snapshot("WebViewContextMenu-01-link")
@@ -78,8 +78,8 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 //
 //        // Image
 //        navigator.openURL("http://wikipedia.org")
-//        waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
-//        waitForExistence(app.webViews.element(boundBy: 0).images.element(boundBy: 0), timeout: 5)
+//        mozWaitForElementToNotExist(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+//        mozWaitForElementToExist(app.webViews.element(boundBy: 0).images.element(boundBy: 0), timeout: 5)
 //        navigator.goto(WebImageContextMenu)
 //        drag()
 //        snapshot("WebViewContextMenu-02-image")
@@ -88,7 +88,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 
     func testWebViewAuthenticationDialog() {
         navigator.openURL("https://jigsaw.w3.org/HTTP/Basic/", waitForLoading: false)
-        waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+        mozWaitForElementToNotExist(app.staticTexts["XCUITests-Runner pasted from Fennec"])
         navigator.nowAt(BasicAuthDialog)
         snapshot("WebViewAuthenticationDialog-01", waitForLoadingIndicator: false)
     }
@@ -96,7 +96,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
     func test3ReloadButtonContextMenu() {
         navigator.openURL(loremIpsumURL)
         waitUntilPageLoad()
-        waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+        mozWaitForElementToNotExist(app.staticTexts["XCUITests-Runner pasted from Fennec"])
 
         navigator.toggleOff(userState.requestDesktopSite, withAction: Action.ToggleRequestDesktopSite)
         navigator.goto(ReloadLongPressMenu)
@@ -108,7 +108,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 
     func testTopSitesMenu() {
         sleep(3)
-        // waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 15)
+        // mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 15)
         navigator.nowAt(NewTabScreen)
         app.collectionViews.cells[AccessibilityIdentifiers.FirefoxHomepage.TopSites.itemCell].firstMatch.swipeUp()
         snapshot("TopSitesMenu-00")
@@ -116,7 +116,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
         // Workaround since in some locales Top Sites are not shown right away
 //        navigator.goto(SettingsScreen)
 //        navigator.goto(HomePanel_TopSites)
-//        waitForExistence(app.buttons["urlBar-cancel"], timeout: 15)
+//        mozWaitForElementToExist(app.buttons["urlBar-cancel"], timeout: 15)
 //        app.buttons["urlBar-cancel"].tap()
 //        navigator.goto(TopSitesPanelContextMenu)
 //        snapshot("TopSitesMenu-01")
@@ -124,14 +124,14 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 
     func testHistoryTableContextMenu() {
         navigator.openURL(loremIpsumURL)
-        waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"], timeoutValue: 5)
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
+        mozWaitForElementToNotExist(app.staticTexts["XCUITests-Runner pasted from Fennec"], timeoutValue: 5)
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
         navigator.performAction(Action.OpenNewTabFromTabTray)
         navigator.performAction(Action.CloseURLBarOpen)
         navigator.nowAt(NewTabScreen)
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
         navigator.goto(LibraryPanel_History)
-        waitForExistence(app.tables["History List"])
+        mozWaitForElementToExist(app.tables["History List"])
         app.tables["History List"].cells.element(boundBy: 1).staticTexts.element(boundBy: 1).press(forDuration: 2)
         snapshot("HistoryTableContextMenu-01")
     }
@@ -140,8 +140,8 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
         sleep(3)
         navigator.openURL(loremIpsumURL)
         // There is no other way the test work with the new Copied.. snackbar ahow on iOS14
-        waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 15)
+        mozWaitForElementToNotExist(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 15)
         navigator.performAction(Action.Bookmark)
         navigator.createNewTab()
         // Disable due to issue #7521
@@ -157,13 +157,13 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
     func test21ReaderModeSettingsMenu() {
         loadWebPage(url: "en.m.wikipedia.org/wiki/Main_Page")
         app.buttons[ AccessibilityIdentifiers.Toolbar.readerModeButton].tap()
-        waitForExistence(app.buttons["ReaderModeBarView.settingsButton"])
+        mozWaitForElementToExist(app.buttons["ReaderModeBarView.settingsButton"])
         app.buttons["ReaderModeBarView.settingsButton"].tap()
         snapshot("21ReaderModeSettingsMenu-01")
     }*/
 
     func testETPperSite() {
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
         navigator.nowAt(NewTabScreen)
         // Enable Strict ETP
         navigator.goto(TrackingProtectionSettings)
@@ -174,8 +174,8 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 
         // Website without blocked elements
         navigator.openURL(loremIpsumURL)
-        waitForNoExistence(app.staticTexts["XCUITests-Runner pasted from Fennec"])
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.trackingProtection], timeout: 5)
+        mozWaitForElementToNotExist(app.staticTexts["XCUITests-Runner pasted from Fennec"])
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.trackingProtection], timeout: 5)
         navigator.goto(TrackingProtectionContextMenuDetails)
         snapshot("TrackingProtectionEnabledPerSite-01")
 
@@ -186,20 +186,20 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
     }
 
     func testSettingsETP() {
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
         navigator.nowAt(NewTabScreen)
         navigator.goto(TrackingProtectionSettings)
 
-        waitForExistence(app.cells["Settings.TrackingProtectionOption.BlockListBasic"])
+        mozWaitForElementToExist(app.cells["Settings.TrackingProtectionOption.BlockListBasic"])
         app.cells["Settings.TrackingProtectionOption.BlockListBasic"].buttons.firstMatch.tap()
         snapshot("TrackingProtectionBasicMoreInfo-01")
 
-        waitForExistence(app.navigationBars["Client.TPAccessoryInfo"])
+        mozWaitForElementToExist(app.navigationBars["Client.TPAccessoryInfo"])
         // Go back to TP settings
         app.navigationBars["Client.TPAccessoryInfo"].buttons.firstMatch.tap()
 
         // See Strict mode info
-        waitForExistence(app.cells["Settings.TrackingProtectionOption.BlockListStrict"])
+        mozWaitForElementToExist(app.cells["Settings.TrackingProtectionOption.BlockListStrict"])
         app.cells["Settings.TrackingProtectionOption.BlockListStrict"].buttons.firstMatch.tap()
         app.tables.cells.staticTexts.firstMatch.swipeUp()
         snapshot("TrackingProtectionStrictMoreInfo-02")
@@ -208,22 +208,22 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
     func testMenuOnTopSites() {
         typealias homeTabBannerA11y = AccessibilityIdentifiers.FirefoxHomepage.HomeTabBanner
 
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
         navigator.nowAt(NewTabScreen)
         navigator.goto(BrowserTabMenu)
         snapshot("MenuOnTopSites-01")
 
         // Set as Default browser screenshot
         navigator.goto(NewTabScreen)
-        waitForExistence(app.buttons[homeTabBannerA11y.ctaButton], timeout: 15)
+        mozWaitForElementToExist(app.buttons[homeTabBannerA11y.ctaButton], timeout: 15)
         app.buttons[homeTabBannerA11y.ctaButton].tap()
-        waitForExistence(app.buttons["HomeTabBanner.goToSettingsButton"], timeout: 15)
+        mozWaitForElementToExist(app.buttons["HomeTabBanner.goToSettingsButton"], timeout: 15)
         snapshot("HomeDefaultBrowserLearnMore")
     }
 
     func testSettings() {
         let table = app.tables.element(boundBy: 0)
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
         navigator.nowAt(NewTabScreen)
         navigator.goto(SettingsScreen)
         table.forEachScreen { i in
@@ -239,14 +239,14 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
     }
 
     func testPrivateBrowsingTabsEmptyState() {
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
         navigator.nowAt(NewTabScreen)
         navigator.toggleOn(userState.isPrivate, withAction: Action.TogglePrivateMode)
         snapshot("PrivateBrowsingTabsEmptyState-01")
     }
 
     func testTakeMarketingScreenshots() {
-        waitForExistence(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
+        mozWaitForElementToExist(app.buttons[AccessibilityIdentifiers.Toolbar.settingsMenuButton], timeout: 10)
         snapshot("00TopSites")
 
         // go to synced tabs home screen
@@ -265,7 +265,7 @@ class L10nSuite1SnapshotTests: L10nBaseSnapshotTests {
 
         // perform a search but don't complete (we're testing autocomplete here)
         navigator.createNewTab()
-        waitForExistence(app.textFields["url"], timeout: 10)
+        mozWaitForElementToExist(app.textFields["url"], timeout: 10)
         app.typeText("firef")
         sleep(2)
         snapshot("01SearchResults")