--- a/Tests/XCUITests/ActivityStreamTest.swift
+++ b/Tests/XCUITests/ActivityStreamTest.swift
@@ -174,7 +174,7 @@ class ActivityStreamTest: BaseTestCase {
         // The website is open
         XCTAssertFalse(TopSiteCellgroup.exists)
         XCTAssertTrue(app.textFields["url"].exists)
-        waitForValueContains(app.textFields["url"], value: "wikipedia.org")
+        mozWaitForValueContains(app.textFields["url"], value: "wikipedia.org")
     }
 
     // Smoketest