--- a/Tests/XCUITests/AuthenticationTest.swift
+++ b/Tests/XCUITests/AuthenticationTest.swift
@@ -8,11 +8,11 @@ let testBasicHTTPAuthURL = "https://jigsaw.w3.org/HTTP/Basic/"
 
 class AuthenticationTest: BaseTestCase {
     func testBasicHTTPAuthenticationPromptVisible() {
-        waitForExistence(app.textFields["url"], timeout: 5)
+        mozWaitForElementToExist(app.textFields["url"], timeout: 5)
         navigator.nowAt(NewTabScreen)
         navigator.openURL(testBasicHTTPAuthURL)
-        waitForExistence(app.staticTexts["Authentication required"], timeout: 100)
-        waitForExistence(app.staticTexts["A username and password are being requested by jigsaw.w3.org. The site says: test"])
+        mozWaitForElementToExist(app.staticTexts["Authentication required"], timeout: 100)
+        mozWaitForElementToExist(app.staticTexts["A username and password are being requested by jigsaw.w3.org. The site says: test"])
 
         let placeholderValueUsername = app.alerts.textFields.element(boundBy: 0).value as! String
         let placeholderValuePassword = app.alerts.secureTextFields.element(boundBy: 0).value as! String
@@ -20,12 +20,12 @@ class AuthenticationTest: BaseTestCase {
         XCTAssertEqual(placeholderValueUsername, "Username")
         XCTAssertEqual(placeholderValuePassword, "Password")
 
-        waitForExistence(app.alerts.buttons["Cancel"])
-        waitForExistence(app.alerts.buttons["Log in"])
+        mozWaitForElementToExist(app.alerts.buttons["Cancel"])
+        mozWaitForElementToExist(app.alerts.buttons["Log in"])
 
         // Skip login due to HTTP Basic Authentication crash in #5757
         // Dismiss authentication prompt
         app.alerts.buttons["Cancel"].tap()
-        waitForNoExistence(app.alerts.buttons["Cancel"], timeoutValue: 5)
+        mozWaitForElementToNotExist(app.alerts.buttons["Cancel"], timeoutValue: 5)
     }
 }