… Tests/XCUITests/BaseTestCase.swift
… Tests/XCUITests/BookmarkingTests.swift
… Tests/XCUITests/BrowsingPDFTests.swift
… Tests/XCUITests/ClipBoardTests.swift
… Tests/XCUITests/DataManagementTests.swift
… Tests/XCUITests/DatabaseFixtureTest.swift
… Tests/XCUITests/DesktopModeTests.swift
… Tests/XCUITests/DisplaySettingsTests.swift
… Tests/XCUITests/DomainAutocompleteTest.swift
… Tests/XCUITests/DownloadFilesTests.swift
… Tests/XCUITests/DragAndDropTests.swift
… Tests/XCUITests/EngagementNotificationTests.swift
… Tests/XCUITests/ExperimentIntegrationTests.swift
… Tests/XCUITests/FindInPageTest.swift
… Tests/XCUITests/FxScreenGraph.swift
… Tests/XCUITests/HistoryTests.swift
… Tests/XCUITests/HomeButtonTests.swift
… Tests/XCUITests/HomePageSettingsUITest.swift
… Tests/XCUITests/IntegrationTests.swift
… Tests/XCUITests/JumpBackInTests.swift
… Tests/XCUITests/LibraryTests.swift
… Tests/XCUITests/MailAppSettingsTests.swift
… Tests/XCUITests/NavigationTest.swift
… Tests/XCUITests/NewTabSettings.swift
… Tests/XCUITests/NightModeTests.swift
… Tests/XCUITests/NoImageTests.swift
… Tests/XCUITests/OnboardingTests.swift
… Tests/XCUITests/OpeningScreenTests.swift
… Tests/XCUITests/PerformanceTests.swift
… Tests/XCUITests/PhotonActionSheetTest.swift