## :scroll: Tickets
[Jira ticket](https://mozilla-hub.atlassian.net/browse/FXIOS-TODO)
[Github issue](https://github.com/mozilla-mobile/firefox-ios/issues/TODO)

## :bulb: Description
The current XCUITest methods waitForExistence, waitForNoExistence, waitForValueContains, and waitFor are very slow-running and need to be optimized.

Performance testing them using measure gave measurment of ~1.5 seconds up to 3 seconds per execution.

Updating those methods to use a simple while-loop takes execution time down to 0.05 up to 0.101, a performance improvement of about 25x for the average element assertion.

The current count of the use of those methods are as follows:

waitForExistence: 642
waitForNoExistence: 72
waitForValueContains: 59
Total: 773

The current total runtime for those assertions is on average 1159.50 seconds.

Replacing those with the updated code will reduce the runtime on average by 695.70 seconds, a runtime savings of over 10 minutes.

The current Full Functional Test Suite runs on average at 3778.881 seconds and will be reduced to an average runtime of 3083.181 seconds, a performance improvement of 28.5%.

## :pencil: Checklist
You have to check all boxes before merging
- [ ] Filled in the above information (tickets numbers and description of your work)
- [ ] Updated the PR name to follow our [PR naming guidelines](https://github.com/mozilla-mobile/firefox-ios/wiki/Pull-Request-Naming-Guide)
- [ ] Wrote unit tests and/or ensured the tests suite is passing
- [ ] When working on UI, I checked and implemented accessibility (minimum Dynamic Text and VoiceOver)
- [ ] If needed I updated documentation / comments for complex code and public methods