@isabelrios 2023-09-15
@jjSDET this is looking really good... 
There are two tests failing consistently on Bitrise:

- testDefaultTopSites
- testTopSitesRemoveAllExceptDefaultClearPrivateData

Not sure if those are fix in a new commit?