@isabelrios 2023-09-15
Even with re-tries for 3 tests, this PR run faster than the regular ones 1-2 minutes. That would increase once the tests are fixed which will improve the build test execution time a bit more. Great work!