@isabelrios 2023-09-18
@jjSDET as we discussed in Slack, the plan would be to split the work here in two PRs. This one to fix all tests except for the ones in the ActivityStream test suite, and then create a follow up PR to update those, is that right?